+++ src/components/Editor/InstrumentGrid/styled.js
