+++ src/components/Editor/InstrumentsManager/styled.js
