+++ src/components/Editor/SamplePlayer/index.js
