+++ src/components/Editor/Sequence/styled.js
