+++ src/components/Editor/SequenceManager/styled.js
