+++ src/components/Editor/SongSettings/index.js
@@ -9,7 +9,13 @@ import {
   Input,
 } from './styled'
 
-const GridSettings = ({ bpm, handleBPM }) => (
+/**
+ * Allows user to set the tempo of the song
+ * @param {number} bpm - The base tempo of the song (comes from Redux)
+ * @param {Function} setBPM - The dispatcher to update the bpm value in our redux store
+ */
+
+const SongSettings = ({ bpm, handleBPM }) => (
   <Menu>
     <Field>BPM: </Field>
     <Input title='Set the current tempo of the song' type='number' value={bpm} onChange={handleBPM} />
@@ -35,4 +41,4 @@ const enhancer: HOC<*, {}> = compose(
   })
 )
 
-export default enhancer(GridSettings)
+export default enhancer(SongSettings)
