+++ src/data/audio/tr909.js
@@ -9,23 +9,6 @@ import tr909highTom from 'assets/samples/tr909/909HighTom.WAV'
 import tr909rimshot from 'assets/samples/tr909/909Rimshot.WAV'
 import tr909snare from 'assets/samples/tr909/909Snare.WAV'
 
-export const tr909Minimal = [
-  {
-    id: `Instrument ${Date.now() + 1}`,
-    name: '909 Kick',
-    samplePath: tr909kick,
-    sampleSource: undefined,
-    sequences: [],
-  },
-  {
-    id: `Instrument ${Date.now() + 9}`,
-    name: '909 Snare',
-    samplePath: tr909snare,
-    sampleSource: undefined,
-    sequences: [],
-  },
-]
-
 export const tr909 = [
   {
     id: `Instrument ${Date.now() + 1}`,
